fix: fix success bug when deleting the same id

+++ src/app/service/TutorService.js
@@ -19,6 +19,10 @@ module.exports = class TutorService {
 
     static async updateTutor(tutorId, tutorData) {
         try {
+            const tutor = await TutorService.getByTutorId(tutorId)
+            if (!tutor) {
+                throw new Error('Tutor not found');
+            }
             await TutorRepository.updateTutor(tutorId, tutorData); 
         } catch (error) {
             throw new Error(error.message);
@@ -27,10 +31,21 @@ module.exports = class TutorService {
     
     static async removeTutor(tutorId) {
         try {
+            const tutor = await TutorService.getByTutorId(tutorId)
+            if (!tutor) {
+                throw new Error('Tutor not found');
+            }
             await TutorRepository.removeTutor(tutorId)
         } catch (error) {
             throw new Error(error.message);
         }  
     }
-   
+    
+    static async getByTutorId(tutorId) {
+        try {
+            return await TutorRepository.getByTutorId(tutorId)
+        } catch (error) {
+            throw new Error(error.message);
+        }
+    }
 };
